requirements.txt
streamlit
python-dotenv
pinecone
google-generativeai
aiohttp
crawl4ai